# docs/swagger.yaml
basePath: /
definitions:
...
  server.swaggerCreatePrescriptionRequest:
    properties:
      appointment_id:
        type: string
      details:
        type: string
      expiration_date:
        type: string
      hospital_id:
        type: string
      patient_id:
        type: string
    required:
    - appointment_id
    - details
    - expiration_date
    - hospital_id
    - patient_id
    type: object
...
  server.swaggerPrescriptionResponse:
    properties:
      appointment_id:
        type: string
      details:
        type: string
      doctor_id:
        type: string
      expiration_date:
        type: string
      hospital_id:
        type: string
      id:
        type: string
      patient_id:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@clinicwise.io
    name: Clinic Wise Support
  description: REST API for the Clinic Wise healthcare management platform.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Clinic Wise API
  version: "1.0"
paths:
...
  /prescriptions:
    post:
      consumes:
      - application/json
      description: Creates a new prescription for a patient. Requires doctor role.
      parameters:
      - description: Create prescription payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server.swaggerCreatePrescriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.swaggerPrescriptionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a prescription
      tags:
      - Prescriptions
  /prescriptions/{id}:
    get:
      description: Returns a single prescription by ID. Requires doctor or patient
        role.
      parameters:
      - description: Prescription ID (ULID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.swaggerPrescriptionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a prescription
      tags:
      - Prescriptions
  /prescriptions/{id}/dispatch:
    patch:
      description: Marks a prescription as dispatched. Requires pharmacist role.
      parameters:
      - description: Prescription ID (ULID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.swaggerPrescriptionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Dispatch a prescription
      tags:
      - Prescriptions
  /prescriptions/appointment/{id}:
    get:
      description: Returns all prescriptions for a given appointment. Requires doctor
        or patient role.
      parameters:
      - description: Appointment ID (ULID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/server.swaggerPrescriptionResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List prescriptions for an appointment
      tags:
      - Prescriptions
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Example: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"